ref: `backgound` and `element-background` should be in different files.
