+++ src/elements/index.ts
@@ -1,3 +1,4 @@
 export * from "./image";
 export * from "./line-break";
 export * from "./paragraph";
+export * from "./background";
