+++ src/index.ts
@@ -1,4 +1,3 @@
-export * from "./background";
 export * from "./base";
 export * from "./elements";
 export * from "./layouts";
